Moved solana scaffold utils to solana package
