+++ packages/scaffold-utils/src/HelpersUtil.ts
@@ -5,11 +5,6 @@ export const HelpersUtil = {
   caipNetworkIdToNumber(caipnetworkId?: CaipNetworkId) {
     return caipnetworkId ? Number(caipnetworkId.split(':')[1]) : undefined
   },
-
-  caipNetworkIdToString(caipnetworkId?: CaipNetworkId) {
-    return caipnetworkId ? String(caipnetworkId.split(':')[1]) : undefined
-  },
-
   getCaipTokens(tokens?: Tokens) {
     if (!tokens) {
       return undefined
