+++ packages/solana/exports/index.ts
@@ -6,6 +6,8 @@ import type { Web3ModalOptions } from '../src/client.js'
 
 export type { Web3ModalOptions } from '../src/client.js'
 
+export { defaultSolanaConfig } from '../src/utils/defaultConfig.js'
+
 export function createWeb3Modal(options: Web3ModalOptions) {
   return new Web3Modal({
     ...options,
