+++ packages/solana/exports/react.tsx
@@ -3,11 +3,12 @@
 import { useSnapshot } from 'valtio'
 import { ConstantsUtil } from '@web3modal/scaffold-utils'
 import { getWeb3Modal } from '@web3modal/scaffold-react'
-import { SolStoreUtil } from '@web3modal/scaffold-utils/solana'
+
+import { SolStoreUtil } from '../src/utils/scaffold'
 import { Web3Modal } from '../src/client.js'
 
 import type { Web3ModalOptions } from '../src/client.js'
-import type { Provider } from '@web3modal/scaffold-utils/solana'
+import type { Provider } from '../src/utils/scaffold'
 
 // -- Setup -------------------------------------------------------------------
 let modal: Web3Modal | undefined = undefined
@@ -63,4 +64,5 @@ export {
   useWeb3ModalEvents
 } from '@web3modal/scaffold-react'
 
+// -- Universal Exports -------------------------------------------------------
 export { defaultSolanaConfig } from '../src/utils/defaultConfig.js'
