--- packages/solana/src/utils/constants.ts
@@ -1,1 +1,0 @@
-export const WALLET_CHOICE_KEY = 'solana.wallet'
