+++ packages/solana/src/utils/favouriteDomain.ts
@@ -1,7 +1,8 @@
 import { PublicKey } from '@solana/web3.js'
 
 import type { Schema } from 'borsh'
-import type { Tag } from '@web3modal/scaffold-utils/solana'
+
+import type { Tag } from './scaffold'
 
 // Class used solely for borsh deserialize function
 export class FavouriteDomain {
