+++ packages/solana/src/utils/scaffold/SolanaConstantsUtil.ts
