+++ packages/solana/src/utils/scaffold/SolanaTypesUtil.ts
